Highscore File Fixed an issue with the highscore file. I created just a basic text file locally to test the tracking and updating of the game highscore but forgot about how this will not work for anyone else if they download it as the file address was hardcoded to my local environement.
New logic will create a new txt file locally on the users desktop automatically and keep it updated.

obviously this isn't a final solution as the file is easily found and can be modified by the user.

Just a quick fix to get the program to run for anyone, just local data so whatever. Eventually will move this to a database and post the scores on my website or something for fun.